requirements.txt
pydantic==1.9.1
docker>=6.0.1
urllib3== 1.26.15
tomli==2.0.1
tomli-w==1.0.0
PyGithub==1.55
click==8.1.3
rich==12.6.0
fsspec==2022.11.0
PyYAML==6.0
jsonschema==4.17.3
colorama==0.4.6
cryptography==39.0.1
kubernetes==23.6.0
pyOpenSSL==23.1.1